Fix filterByVersion method
