+++ lib/common/versionUtils.js
@@ -4,8 +4,7 @@ const VERSION_30 = { key: 'openapi', version: '3.0' },
   GENERIC_VERSION2 = { key: 'swagger', version: '2.' },
   GENERIC_VERSION3 = { key: 'openapi', version: '3.' },
   DEFAULT_SPEC_VERSION = VERSION_30.version,
-  fs = require('fs'),
-  _ = require('lodash');
+  fs = require('fs');
 
 /**
  * gets the version key and the version and generates a regular expression that
@@ -132,26 +131,15 @@ function getConcreteSchemaUtils({ type, data }) {
 
 /**
    * Filter the options by the version provided by the user
-   * @param {array} optsArray The options to be filtered
+   * @param {array} options The options to be filtered
    * @param {string} version The version that will be used
    * @returns {array} the filtered options related to the version used
    */
-function filterOptionsByVersion(optsArray, version) {
-  let defOptions = {};
-  optsArray = optsArray.filter((option) => {
+function filterOptionsByVersion(options, version) {
+  options = options.filter((option) => {
     return option.supportedIn.includes(version);
   });
-  _.each(optsArray, (opt) => {
-    // special handling for indent character as in documentation it states `Tab` and `Space`
-    // but for the generation mode, we need actual values
-    if (opt.id === 'indentCharacter') {
-      defOptions[opt.id] = opt.default === 'tab' ? '\t' : ' ';
-    }
-    else {
-      defOptions[opt.id] = opt.default;
-    }
-  });
-  return defOptions;
+  return options;
 }
 
 module.exports = {
