+++ lib/options.js
@@ -69,7 +69,7 @@ module.exports = {
         ' values: `description`, `operationid`, `url`.',
         external: true,
         usage: ['CONVERSION', 'VALIDATION'],
-        supportedIn: [VERSION31]
+        supportedIn: [VERSION30, VERSION31]
       },
       {
         name: 'Set indent character',
@@ -299,7 +299,19 @@ module.exports = {
 
     if (mode === 'use') {
       // options to be used as default kv-pairs
-      return filterOptionsByVersion(optsArray, version);
+      let defOptions = {};
+      optsArray = filterOptionsByVersion(optsArray, version);
+      _.each(optsArray, (opt) => {
+        // special handling for indent character as in documentation it states `Tab` and `Space`
+        // but for the generation mode, we need actual values
+        if (opt.id === 'indentCharacter') {
+          defOptions[opt.id] = opt.default === 'tab' ? '\t' : ' ';
+        }
+        else {
+          defOptions[opt.id] = opt.default;
+        }
+      });
+      return defOptions;
     }
 
     // options to be used as documentation
